+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 0.0.26 – More Multi-file context implementation
 
-- ADD: First feature for agenting processing - edit file from the workspace
+- ADD: Agent framework + editFile tool (EXPERIMENTAL! Very early implementation so use at own risk) - Allow LLM to edit files in the workspace.
+- ADD: Capabilities framework, UI section
+- FIX: New session will always load 'fresh' files in the context. Also updated how/what files are loaded on new session based on previous context and what file is active when the new session was started.
 
 ## 0.0.25 – More Multi-file context implementation